rev-list: remove 'commit <hash>' line from custom formats
Without this, a command such as:

  $ git rev-list -1 HEAD --pretty=format:%h

Results in the output:

  commit 18f6780b056ad7835a09e37a9bfbede899bfa19b
  18f6780

This behavior is different from a similarly formatted 'log' or 'show'
command, in that the latter only show exactly the format requested by
the user.

Allow rev-list to account for CMIT_FMT_USERFORMAT, and in those cases,
elide printing the 'commit <hash>' line.

Signed-off-by: Vishal Verma <vishal@stellar.sh>